+++ nova-app/src/features/login/ui/LoginButton.tsx
@@ -1,8 +1,8 @@
 import { cn } from '@/shared/utils/cn';
 import { cva } from 'class-variance-authority';
-import KakaoIcon from '@/shared/assets/KakaoTalkIcon.svg';
-import GoogleIcon from '@/shared/assets/GoogleIcon.svg';
-import GithubIcon from '@/shared/assets/GithubIcon.svg';
+import { KakaoIcon } from '@/shared/assets';
+import { GoogleIcon } from '@/shared/assets';
+import { GithubIcon } from '@/shared/assets';
 
 const LoginButtonVariants = cva(
   'w-75 h-11 rounded-interactive-default py-padding-regular px-padding-bold typo-body-key flex flex-row items-center justify-center',
