+++ nova-app/src/shared/assets/index.ts
@@ -1,0 +1,4 @@
+export { default as KakaoIcon } from './KakaoTalkIcon.svg';
+export { default as GoogleIcon } from './GoogleIcon.svg';
+export { default as GithubIcon } from './GithubIcon.svg';
+export { default as NovaLabel } from './NobalLabel.svg';
